Add `BibPage` test, test for correct tab titles
