+++ test/unit/BibDetails.test.js
@@ -5,289 +5,7 @@ import { shallow, mount } from 'enzyme';
 
 // Import the component that is going to be tested
 import BibDetails from './../../src/app/components/BibPage/BibDetails';
-
-const bibs = [
-  {
-    '@type': ['nypl:Item', 'nypl:Resource'],
-    '@id': 'res:b11417539',
-    carrierType: [{ '@id': 'carriertypes:nc', prefLabel: 'volume' }],
-    createdString: ['1991'],
-    createdYear: 1991,
-    creatorLiteral: ['De Grazia, Margreta.'],
-    dateStartYear: 1991,
-    dateString: ['1991'],
-    dimensions: ['23 cm.'],
-    extent: ['xi, 244 p., [8] p. of plates : ill. ;'],
-    idIsbn: ['0198117787'],
-    idLccn: ['90048121'],
-    idOclc: ['22452233'],
-    identifier: [
-      {
-        '@type': 'bf:ShelfMark',
-        '@value': 'JFD 91-4064',
-      },
-      {
-        '@type': 'nypl:Bnumber',
-        '@value': '11417539',
-      },
-      {
-        '@type': 'bf:Isbn',
-        '@value': '0198117787',
-      },
-      {
-        '@type': 'bf:Isbn',
-        '@value': '012345',
-        identifierStatus: 'incorrect',
-      },
-      {
-        '@type': 'bf:Lccn',
-        '@value': '90048121',
-      },
-      {
-        '@type': 'nypl:Oclc',
-        '@value': '22452233',
-      },
-    ],
-    issuance: [{ '@id': 'urn:biblevel:m', prefLabel: 'monograph/item' }],
-    items: [
-      {
-        '@id': 'res:i13183119',
-        accessMessage: [
-          {
-            '@id': 'accessMessage:1',
-            prefLabel: 'USE IN LIBRARY',
-          },
-        ],
-        catalogItemType: [
-          {
-            '@id': 'catalogItemType:55',
-            prefLabel: 'book, limited circ, MaRLI',
-          },
-        ],
-        holdingLocation: [
-          {
-            '@id': 'loc:mal92',
-            prefLabel: 'Schwarzman Building M2 - General Research Room 315',
-          },
-        ],
-        idBarcode: [
-          '33433040712188',
-        ],
-        identifier: [
-          {
-            '@type': 'bf:ShelfMark',
-            '@value': 'JFD 91-4064',
-          },
-          {
-            '@type': 'bf:Barcode',
-            '@value': '33433040712188',
-          },
-        ],
-        requestable: [
-          false,
-        ],
-        shelfMark: [
-          'JFD 91-4064',
-        ],
-        status: [
-          {
-            '@id': 'status:a',
-            prefLabel: 'Available ',
-          },
-        ],
-        uri: 'i13183119',
-        idNyplSourceId: {
-          '@type': 'SierraNypl',
-          '@value': '13183119',
-        },
-      },
-    ],
-    language: [{ '@id': 'lang:eng', prefLabel: 'English' }],
-    lccClassification: ['PR3071 .D4 1991'],
-    materialType: [{ '@id': 'resourcetypes:txt', prefLabel: 'Text' }],
-    mediaType: [{ '@id': 'mediatypes:n', prefLabel: 'unmediated' }],
-    note: [
-      {
-        noteType: 'Bibliography',
-        prefLabel: 'Includes bibliographical references (p. [227]-238) and index.',
-        '@type': 'bf:Note',
-      },
-    ],
-    numAvailable: 1,
-    numItems: 1,
-    placeOfPublication: ['Oxford [England] : New York :'],
-    publicationStatement: ['Oxford [England] : Clarendon Press ; New York : Oxford University Press, 1991.'],
-    publisherLiteral: ['Clarendon Press ; Oxford University Press,'],
-    shelfMark: ['JFD 91-4064'],
-    subjectHeadingData:[
-      {
-        "label": "Shakespeare, William, -- 1564-1616",
-        "uuid": "56567c13-9c92-4379-8f8a-7e6168ef47c1",
-        "bib_count": 19632,
-        "desc_count": 2024,
-        "parent": {
-          "label": "Shakespeare, William,",
-          "uuid": "584104d8-9be1-45ba-bfd7-b59f17d36c88",
-          "bib_count": 19642,
-          "desc_count": 2042
-        }
-      },
-      {
-        "label": "Malone, Edmond, -- 1741-1812",
-        "uuid": "1cec52b3-dfe0-44ab-9cfb-4be9d4d0b97e",
-        "bib_count": 26,
-        "desc_count": 7,
-        "parent": {
-          "label": "Malone, Edmond,",
-          "uuid": "c8c76c90-3d01-49b0-90ca-7204c54faedf",
-          "bib_count": 26,
-          "desc_count": 8
-        }
-      },
-      {
-        "label": "Shakespeare, William, -- 1564-1616 -- Criticism, Textual",
-        "uuid": "1913e567-8d33-41d9-9a8b-104008655684",
-        "bib_count": 502,
-        "desc_count": 4,
-        "parent": {
-          "label": "Shakespeare, William, -- 1564-1616",
-          "uuid": "56567c13-9c92-4379-8f8a-7e6168ef47c1",
-          "bib_count": 19632,
-          "desc_count": 2024,
-          "parent": {
-            "label": "Shakespeare, William,",
-            "uuid": "584104d8-9be1-45ba-bfd7-b59f17d36c88",
-            "bib_count": 19642,
-            "desc_count": 2042
-          }
-        }
-      },
-      {
-        "label": "Editing -- History -- 18th century",
-        "uuid": "861d31ed-ec55-44d3-ac7e-9b3b77826304",
-        "bib_count": 13,
-        "desc_count": 1,
-        "parent": {
-          "label": "Editing -- History",
-          "uuid": "c9619cd0-105a-41db-93ae-70bd2fdea6ea",
-          "bib_count": 49,
-          "desc_count": 12,
-          "parent": {
-            "label": "Editing",
-            "uuid": "39916e16-51b2-45d2-83b2-d2d9f2b76da3",
-            "bib_count": 702,
-            "desc_count": 69
-          }
-        }
-      }
-    ],
-    title: ['Shakespeare verbatim : the reproduction of authenticity and the 1790 apparatus'],
-    titleDisplay: ['Shakespeare verbatim : the reproduction of authenticity and the 1790 apparatus / Margreta de Grazia.'],
-    type: ['nypl:Item'],
-    updatedAt: 1524665261653,
-    uri: 'b11417539',
-    suppressed: false,
-  },
-  {
-    '@type': ['nypl:Item', 'nypl:Resource'],
-    '@id': 'res:b11417539',
-    carrierType: [{ '@id': 'carriertypes:nc', prefLabel: 'volume' }],
-    createdString: ['1991'],
-    createdYear: 1991,
-    creatorLiteral: ['De Grazia, Margreta.'],
-    dateStartYear: 1991,
-    dateString: ['1991'],
-    dimensions: ['23 cm.'],
-    extent: ['xi, 244 p., [8] p. of plates : ill. ;'],
-    idIsbn: ['0198117787'],
-    idLccn: ['90048121'],
-    idOclc: ['22452233'],
-    identifier: [
-      'urn:callnumber:JFD 91-4064',
-      'urn:bnum:11417539',
-      'urn:isbn:0198117787',
-      'urn:isbn:012345',
-      'urn:lccn:90048121',
-      'urn:oclc:22452233',
-    ],
-    issuance: [{ '@id': 'urn:biblevel:m', prefLabel: 'monograph/item' }],
-    items: [
-      {
-        '@id': 'res:i13183119',
-        accessMessage: [
-          {
-            '@id': 'accessMessage:1',
-            prefLabel: 'USE IN LIBRARY',
-          },
-        ],
-        catalogItemType: [
-          {
-            '@id': 'catalogItemType:55',
-            prefLabel: 'book, limited circ, MaRLI',
-          },
-        ],
-        holdingLocation: [
-          {
-            '@id': 'loc:mal92',
-            prefLabel: 'Schwarzman Building M2 - General Research Room 315',
-          },
-        ],
-        idBarcode: [
-          '33433040712188',
-        ],
-        identifier: [
-          'urn:callnumber:JFD 91-4064',
-          'urn:barcode:33433040712188',
-        ],
-        requestable: [
-          false,
-        ],
-        shelfMark: [
-          'JFD 91-4064',
-        ],
-        status: [
-          {
-            '@id': 'status:a',
-            prefLabel: 'Available ',
-          },
-        ],
-        uri: 'i13183119',
-        idNyplSourceId: {
-          '@type': 'SierraNypl',
-          '@value': '13183119',
-        },
-      },
-    ],
-    language: [{ '@id': 'lang:eng', prefLabel: 'English' }],
-    lccClassification: ['PR3071 .D4 1991'],
-    materialType: [{ '@id': 'resourcetypes:txt', prefLabel: 'Text' }],
-    mediaType: [{ '@id': 'mediatypes:n', prefLabel: 'unmediated' }],
-    note: [
-      {
-        noteType: 'Bibliography',
-        prefLabel: 'Includes bibliographical references (p. [227]-238) and index.',
-        '@type': 'bf:Note',
-      },
-    ],
-    numAvailable: 1,
-    numItems: 1,
-    placeOfPublication: ['Oxford [England] : New York :'],
-    publicationStatement: ['Oxford [England] : Clarendon Press ; New York : Oxford University Press, 1991.'],
-    publisherLiteral: ['Clarendon Press ; Oxford University Press,'],
-    shelfMark: ['JFD 91-4064'],
-    subjectLiteral:
-     ['Editing -- History -- 18th century.',
-       'Malone, Edmond, 1741-1812.',
-       'Shakespeare, William, 1564-1616 -- Criticism, Textual.',
-       'Shakespeare, William, 1564-1616.'],
-    title: ['Shakespeare verbatim : the reproduction of authenticity and the 1790 apparatus'],
-    titleDisplay: ['Shakespeare verbatim : the reproduction of authenticity and the 1790 apparatus / Margreta de Grazia.'],
-    type: ['nypl:Item'],
-    updatedAt: 1524665261653,
-    uri: 'b11417539',
-    suppressed: false,
-  },
-];
+import bibs from '../fixtures/bibs';
 
 describe('BibDetails', () => {
   describe('Invalid props', () => {
